changed spelling,grammers,uppercase,lowercase.
I had updated the mistakes in the Readme file, and now according to me the file is free from all types of mistakes. In some area i didn't get the text or what you want to say, so about this context will we discuss later on. Please review the file and merge my pull request.